pyproject.toml
[project]
name = "okprintdesigns"
version = "0.1.0"
description = "Local AI-powered print design system for vector graphics and style transfer"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "okprintdesigns" }]
dependencies = [
    "torch>=2.1.0",
    "diffusers>=0.27.0",
    "transformers>=4.40.0",
    "accelerate>=0.29.0",
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "huggingface-hub>=0.22.0",
    "peft>=0.10.0",
    "controlnet-aux>=0.0.7",
    "invisible-watermark>=0.2.0",
    "sentencepiece>=0.2.0",
    " optimum-quanto>=0.2.0; sys_platform == 'linux'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
cuda = [
    "torch>=2.1.0+cu121",
]
cpu = [
    "torch>=2.1.0+cpu",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
